requirements.txt
tensorflow-cpu
flask
nltk
gunicorn

